refetch templates every 1 minute
