+++ src/context/TemplatesProvider/TemplatesProviderContext.js
@@ -1,32 +1,16 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
-import { akashlyticsApi } from "../../shared/constants";
+import React from "react";
+import { useTemplates as useTemplatesQuery } from "../../queries/useTemplatesQuery";
 
 const TemplatesProviderContext = React.createContext({});
 
 export const TemplatesProvider = ({ children }) => {
-  const [categories, setCategories] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
+  const { data, isFetching: isLoading } = useTemplatesQuery();
+  const categories = data ? data.categories : [];
+  const templates = data ? data.templates : [];
 
   function getTemplateByPath(path) {
     return categories.flatMap((x) => x.templates).find((x) => x.path === path);
   }
-
-  const templates = categories.flatMap((x) => x.templates);
-
-  useEffect(() => {
-    (async () => {
-      setIsLoading(true);
-
-      const response = await axios.get(`${akashlyticsApi}/templates`);
-      let categories = response.data.filter((x) => (x.templates || []).length > 0);
-      categories.forEach((c) => {
-        c.templates.forEach((t) => (t.category = c.title));
-      });
-      setCategories(categories);
-      setIsLoading(false);
-    })();
-  }, []);
 
   return <TemplatesProviderContext.Provider value={{ isLoading, categories, templates, getTemplateByPath }}>{children}</TemplatesProviderContext.Provider>;
 };
